feat: body sanitizer
