+++ src/middleware/request-logger.js
@@ -1,17 +1,6 @@
 const morgan = require('morgan');
 const logger = require('../config/logger');
-const { evolve, reduce, always } = require('ramda');
-
-const sensitiveKeyList = ['password'];
-const transformations = reduce(
-  (acc, key) => ({
-    ...acc,
-    [key]: always('********'),
-  }),
-  {},
-  sensitiveKeyList
-);
-const bodySanitizer = evolve(transformations);
+const { bodySanitizer } = require('./helpers');
 
 const requestLogger = morgan((token, req) => {
   const body = req.body;
